machine_learning_model/requirements.txt
absl-py==1.3.0
anyio==3.6.2
astunparse==1.6.3
cached-property==1.5.2
cachetools==5.2.0
charset-normalizer==2.1.1
classification-models-3D==1.0.3
click==8.1.3
colorama==0.4.6
cycler==0.11.0
efficientnet-3D==1.0.2
fastapi==0.87.0
flatbuffers==22.10.26
fonttools==4.38.0
gast==0.4.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
gunicorn==20.1.0
h11==0.14.0
h5py==3.1.0
idna==3.4
importlib-metadata==5.0.0
keras==2.11.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
numpy==1.21.6
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
Pillow==9.3.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.5
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
segmentation-models-3D==1.0.2
six==1.15.0
sniffio==1.3.0
starlette==0.21.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
tensorflow-serving-api==2.5.0
termcolor==1.1.0
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.20.0
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.12.1
zipp==3.10.0
